spring producer service view added

+++ Spring Cloud Services/employee-consumer/src/main/java/com/java/spring/consumer/controllers/model/DomainEmployee.java
@@ -10,7 +10,7 @@ public class DomainEmployee {
     private String name;
     private String designation;
     private List<String> roles;
-    private double salary;
+    private int salary;
     private int exp;
 
     public String getEmpId() {
@@ -49,11 +49,11 @@ public class DomainEmployee {
         return this;
     }
 
-    public double getSalary() {
+    public int getSalary() {
         return salary;
     }
 
-    public DomainEmployee setSalary(double salary) {
+    public DomainEmployee setSalary(int salary) {
         this.salary = salary;
         return this;
     }
